refactor: tweak package structure
